Configuring tokens, creating custom parameters, validating actions, and securing endpoints

+++ src/tasks/tasks.service.ts
@@ -4,6 +4,7 @@ import { UpdateTaskDto } from 'src/tasks/dto/update-task.dto';
 import { PrismaService } from '../prisma/prisma.service'
 import { CreateTaskDto } from './dto/create-task.dto';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
+import { PlayloadTokenDto } from 'src/users/dto/playload-token.dto';
 
 @Injectable()
 export class TasksService {
@@ -43,14 +44,14 @@ export class TasksService {
     throw new HttpException ("Essa tarefa não existe", HttpStatus.NOT_FOUND)
   }
 
-  async create(createTaskDto: CreateTaskDto){
+  async create(createTaskDto: CreateTaskDto, tokenPlayload: PlayloadTokenDto){
     try{
       const newTask = await this.prisma.task.create({
       data: {
         name: createTaskDto.name,
         description: createTaskDto.description,
         completed: false,
-        userId: createTaskDto.userId,
+        userId: tokenPlayload.sub
       } 
     })
     return newTask;
@@ -60,7 +61,7 @@ export class TasksService {
     }
   }
 
-  async update(id: number, updateTaskDto: UpdateTaskDto){
+  async update(id: number, updateTaskDto: UpdateTaskDto, tokenPlayload: PlayloadTokenDto){
     try{
       const findTask = await this.prisma.task.findFirst({
     where: {
@@ -69,6 +70,10 @@ export class TasksService {
     })
 
     if (!findTask){
+      throw new HttpException ("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
+    }
+
+    if (findTask.userId !== tokenPlayload.sub){
       throw new HttpException ("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
     }
 
@@ -89,7 +94,7 @@ export class TasksService {
     }
     }
 
-    async delete (id: number){
+    async delete (id: number, tokenPlayload: PlayloadTokenDto){
     try{
          const findTask = await this.prisma.task.findFirst({
     where: {
@@ -99,6 +104,10 @@ export class TasksService {
 
     if (!findTask){
       throw new HttpException ("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
+    }
+
+    if (findTask.userId !== tokenPlayload.sub){
+      throw new HttpException ("Falha ao deletar essa tarefa!", HttpStatus.BAD_REQUEST)
     }
 
     await this.prisma.task.delete({
